modify: relocation of files
