--- src/components/features/api/auth/useLogin.ts
@@ -1,24 +1,0 @@
-import { client } from "@/lib/rpc";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import type { InferRequestType, InferResponseType } from "hono";
-import { useRouter } from "next/navigation";
-
-type ResponseType = InferResponseType<(typeof client.api.auth.login)["$post"]>;
-type RequestType = InferRequestType<(typeof client.api.auth.login)["$post"]>;
-
-export const useLogin = () => {
-	const queryClient = useQueryClient();
-	const router = useRouter();
-	const mutation = useMutation<ResponseType, Error, RequestType>({
-		mutationFn: async ({ json }) => {
-			const response = await client.api.auth.login.$post({ json });
-			return await response.json();
-		},
-		onSuccess: () => {
-			router.refresh();
-			queryClient.invalidateQueries({ queryKey: ["currentUser"] });
-		},
-	});
-
-	return mutation;
-};
