--- src/components/features/api/auth/useLogout.ts
@@ -1,22 +1,0 @@
-import { client } from "@/lib/rpc";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import type { InferResponseType } from "hono";
-import { useRouter } from "next/navigation";
-
-type ResponseType = InferResponseType<(typeof client.api.auth.logout)["$post"]>;
-
-export const useLogout = () => {
-	const queryClient = useQueryClient();
-	const router = useRouter();
-	const mutation = useMutation<ResponseType, Error>({
-		mutationFn: async () => {
-			const response = await client.api.auth.logout.$post();
-			return await response.json();
-		},
-		onSuccess: () => {
-			router.refresh();
-			queryClient.invalidateQueries({ queryKey: ["currentUser"] });
-		},
-	});
-	return mutation;
-};
