--- src/components/features/api/auth/useRegister.ts
@@ -1,26 +1,0 @@
-import { client } from "@/lib/rpc";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import type { InferRequestType, InferResponseType } from "hono";
-import { useRouter } from "next/navigation";
-
-type ResponseType = InferResponseType<
-	(typeof client.api.auth.register)["$post"]
->;
-type RequestType = InferRequestType<(typeof client.api.auth.register)["$post"]>;
-
-export const useRegister = () => {
-	const router = useRouter();
-	const queryClient = useQueryClient();
-	const mutation = useMutation<ResponseType, Error, RequestType>({
-		mutationFn: async ({ json }) => {
-			const response = await client.api.auth.register.$post({ json });
-			return await response.json();
-		},
-		onSuccess: () => {
-			router.refresh();
-			queryClient.invalidateQueries({ queryKey: ["currentUser"] });
-		},
-	});
-
-	return mutation;
-};
