+++ src/components/pages/auth/SignUp.tsx
@@ -23,8 +23,8 @@ import { useForm } from "react-hook-form";
 import { FaGithub } from "react-icons/fa";
 import { FcGoogle } from "react-icons/fc";
 import type { z } from "zod";
-import { useRegister } from "../../features/api/auth/useRegister";
-import { signUpSchema } from "../../features/schemas/sign-up-schema";
+import { useRegister } from "../../features/server/auth/api/auth/useRegister";
+import { signUpSchema } from "../../features/server/auth/schemas/sign-up-schema";
 
 function SignUp() {
 	const { mutate } = useRegister();
